Creating a new file lib/features/diary/models/appointment.dart containing the Appointment model class
